fix: error

+++ src/infra/api/index.tsx
@@ -11,7 +11,7 @@ api.interceptors.response.use(
     return response;
   },
   async function (error) {
-    if (error.response.status === 401) {
+    if (error.response && error.response?.status === 401) {
       await AsyncStorage.clear();
       return router.push("/");
     }
